+++ commerce/payment/src/main/java/ru/yandex/practicum/service/PaymentService.java
@@ -3,16 +3,17 @@ package ru.yandex.practicum.service;
 import ru.yandex.practicum.dto.order.OrderDto;
 import ru.yandex.practicum.dto.payment.PaymentDto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface PaymentService {
     PaymentDto createPayment(OrderDto orderDto);
 
-    Double totalCost(OrderDto orderDto);
+    BigDecimal totalCost(OrderDto orderDto);
 
     void processSuccessfulPayment(UUID paymentId);
 
-    Double productCost(OrderDto orderDto);
+    BigDecimal productCost(OrderDto orderDto);
 
     void failedPayment(UUID paymentId);
 }
